+++ src/test/suite/testFiles/truncateString.tsx
@@ -1,0 +1,8 @@
+import { useTranslations } from "@vocab/react";
+import translations from "./.vocab";
+import React from "react";
+
+const MyComponent = () => {
+  const { t } = useTranslations(translations);
+  return <div>This is a long line of text that will be truncated.</div>;
+};